Placing props and adding new panels and lamps
As title suggests, placed around a few of NotNightRay's props and added a few of my more misc ones such as lamps and panels. Other changes include changes in lighting and placement of doorways do to the new props, added a few pillars with metal beams and lit the vents up.